franka_buttons/requirements.txt
websockets>=14
requests>=2
python-dotenv